backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
TensorFlow==2.19.0
OpenCV-Python==4.8.1.78
numpy==1.24.3
Werkzeug==3.0.1
Pillow==10.1.0
python-dotenv==1.0.0